app/requirements.txt
aniso8601==6.0.0
attrs==19.1.0
Click==7.0
Flask==1.0.2
flask-restplus==0.12.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
py-healthcheck==1.8.0
pyrsistent==0.15.1
pytz==2019.1
six==1.12.0
Werkzeug==0.15.2
